+++ rtkbase_update.sh
@@ -154,6 +154,16 @@ upd_2.3.2() {
   systemctl daemon-reload
 }
 
+upd_2.3.2() {
+#Add restart condition in gpsd.service
+  sed -i '/^ExecStart=.*/a Restart=always' /etc/systemd/system/gpsd.service
+  sed -i '/^Restart=always.*/a RestartSec=30' /etc/systemd/system/gpsd.service
+ #update python module
+  python3 -m pip install -r ${destination_directory}'/web_app/requirements.txt' --extra-index-url https://www.piwheels.org/simple
+
+  return
+}
+
 # standard update
 update
 # calling specific update function. If we are using v2.2.5, it will call the function upd_2.2.5
@@ -170,14 +180,15 @@ sed -i 's/^version=.*/'${new_release}'/' ${destination_directory}/settings.conf
 #change rtkbase's content owner
 chown -R ${standard_user}:${standard_user} ${destination_directory}
 
+
 echo 'restart ntrip/rtcm to send the new release number in the stream'
 systemctl is-active --quiet str2str_ntrip.service && systemctl restart str2str_ntrip.service
 systemctl is-active --quiet str2str_local_ntrip_caster.service && systemctl restart str2str_local_ntrip_caster.service
 systemctl is-active --quiet str2str_rtcm_svr.service && systemctl restart str2str_rtcm_svr.service
 systemctl is-active --quiet str2str_rtcm_serial.service && systemctl restart str2str_rtcm_serial.service
 
-echo "Restart web server"
-systemctl restart rtkbase_web.service
-
 #if a reboot is needed
 #systemctl reboot
+
+echo "Restart web server"
+systemctl restart rtkbase_web.service