+++ web_app/RTKLIB.py
@@ -698,11 +698,17 @@ class RTKLIB:
     # and emits them to the connected browser as messages
     def broadcastSatellites(self):
         count = 0
+        last_gps_timestamp=""
 
         while self.server_not_interrupted:
 
             # update satellite levels
             self.rtkc.getObs()
+            #check if gps time is the same (no signal input)
+            if last_gps_timestamp == self.rtkc.obs_rover.get("gps_time"):
+                self.rtkc.obs_rover = {}
+            else:
+                last_gps_timestamp = self.rtkc.obs_rover.get("gps_time")
 
 #            if count % 10 == 0:
             #print("Sending sat rover levels:\n" + str(self.rtkc.obs_rover))
@@ -718,11 +724,17 @@ class RTKLIB:
     # this function reads current rtklib status, coordinates and obs count
     def broadcastCoordinates(self):
         count = 0
-
+        last_receiver_timestamp = ""
         while self.server_not_interrupted:
 
             # update RTKLIB status
             self.rtkc.getStatus()
+
+            #check if receiver time is the same (no signal input)
+            if last_receiver_timestamp == self.rtkc.status.get("time of receiver clock rover"):
+                self.rtkc.status = {}
+            else:
+                last_receiver_timestamp = self.rtkc.status.get("time of receiver clock rover")
 
 #            if count % 10 == 0:
 #                print("Sending RTKLIB status select information:")