+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## [2.3.3] - 2022-02-28
 ### Fixed
 - Fix the Rinex conversion failure. #206
-
+- Restart Ntrip/Rtcm services after an update. #171
 ## [2.3.2] - 2022-02-22
 ### Added
 - GUI -> Status: Added a new default map background : Osm "standard", from osm.org.